[AND-3385] AdMob Open Bidding implementation
