+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/google/ads/mediation/vungle/VungleBaseAdapter.java
@@ -1,0 +1,146 @@
+package com.google.ads.mediation.vungle;
+
+import android.content.Context;
+import android.os.Bundle;
+import android.text.TextUtils;
+import android.util.Log;
+import androidx.annotation.NonNull;
+import com.google.ads.mediation.vungle.VungleInitializer.VungleInitializationListener;
+import com.google.android.gms.ads.mediation.InitializationCompleteCallback;
+import com.google.android.gms.ads.mediation.MediationAdLoadCallback;
+import com.google.android.gms.ads.mediation.MediationConfiguration;
+import com.google.android.gms.ads.mediation.MediationInterstitialAd;
+import com.google.android.gms.ads.mediation.MediationInterstitialAdCallback;
+import com.google.android.gms.ads.mediation.MediationInterstitialAdConfiguration;
+import com.google.android.gms.ads.mediation.VersionInfo;
+import com.google.android.gms.ads.mediation.rtb.RtbAdapter;
+import com.google.android.gms.ads.mediation.rtb.RtbSignalData;
+import com.google.android.gms.ads.mediation.rtb.SignalCallbacks;
+import com.vungle.mediation.BuildConfig;
+import com.vungle.warren.Vungle;
+import java.util.HashSet;
+import java.util.List;
+
+public abstract class VungleBaseAdapter extends RtbAdapter {
+
+  private static final String TAG = VungleBaseAdapter.class.getSimpleName();
+  protected static final String KEY_APP_ID = "appid";
+
+  @Override
+  public void collectSignals(@NonNull RtbSignalData rtbSignalData,
+      @NonNull SignalCallbacks signalCallbacks) {
+    String token = Vungle.getAvailableBidTokens(rtbSignalData.getContext());
+    Log.d(TAG, "token=" + token);
+    signalCallbacks.onSuccess(token);
+  }
+
+  @Override
+  @NonNull
+  public VersionInfo getVersionInfo() {
+    String versionString = BuildConfig.ADAPTER_VERSION;
+    String[] splits = versionString.split("\\.");
+
+    if (splits.length >= 4) {
+      int major = Integer.parseInt(splits[0]);
+      int minor = Integer.parseInt(splits[1]);
+      int micro = Integer.parseInt(splits[2]) * 100 + Integer.parseInt(splits[3]);
+      return new VersionInfo(major, minor, micro);
+    }
+
+    String logMessage =
+        String.format(
+            "Unexpected adapter version format: %s. Returning 0.0.0 for adapter version.",
+            versionString);
+    Log.w(TAG, logMessage);
+    return new VersionInfo(0, 0, 0);
+  }
+
+  @Override
+  @NonNull
+  public VersionInfo getSDKVersionInfo() {
+    String versionString = com.vungle.warren.BuildConfig.VERSION_NAME;
+    String[] splits = versionString.split("\\.");
+
+    if (splits.length >= 3) {
+      int major = Integer.parseInt(splits[0]);
+      int minor = Integer.parseInt(splits[1]);
+      int micro = Integer.parseInt(splits[2]);
+      return new VersionInfo(major, minor, micro);
+    }
+
+    String logMessage =
+        String.format(
+            "Unexpected SDK version format: %s. Returning 0.0.0 for SDK version.", versionString);
+    Log.w(TAG, logMessage);
+    return new VersionInfo(0, 0, 0);
+  }
+
+  @Override
+  public void initialize(
+      @NonNull Context context,
+      @NonNull final InitializationCompleteCallback initializationCompleteCallback,
+      @NonNull List<MediationConfiguration> mediationConfigurations) {
+
+    if (Vungle.isInitialized()) {
+      initializationCompleteCallback.onInitializationSucceeded();
+      return;
+    }
+
+    HashSet<String> appIDs = new HashSet<>();
+    for (MediationConfiguration configuration : mediationConfigurations) {
+      Bundle serverParameters = configuration.getServerParameters();
+      String appIDFromServer = serverParameters.getString(KEY_APP_ID);
+
+      if (!TextUtils.isEmpty(appIDFromServer)) {
+        appIDs.add(appIDFromServer);
+      }
+    }
+
+    int count = appIDs.size();
+    if (count <= 0) {
+      initializationCompleteCallback.onInitializationFailed(
+          "Initialization failed: Missing or Invalid App ID.");
+      return;
+    }
+
+    String appID = appIDs.iterator().next();
+    if (count > 1) {
+      String logMessage =
+          String.format(
+              "Multiple '%s' entries found: %s. Using '%s' to initialize the Vungle SDK.",
+              KEY_APP_ID, appIDs.toString(), appID);
+      Log.w(TAG, logMessage);
+    }
+
+    VungleInitializer.getInstance()
+        .initialize(
+            appID,
+            context.getApplicationContext(),
+            new VungleInitializationListener() {
+              @Override
+              public void onInitializeSuccess() {
+                initializationCompleteCallback.onInitializationSucceeded();
+              }
+
+              @Override
+              public void onInitializeError(String errorMessage) {
+                initializationCompleteCallback.onInitializationFailed(
+                    "Initialization Failed: " + errorMessage);
+              }
+            });
+  }
+
+  @Override
+  public void loadRtbInterstitialAd(
+      @NonNull MediationInterstitialAdConfiguration mediationInterstitialAdConfiguration,
+      @NonNull MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback> mediationAdLoadCallback) {
+    Log.d(TAG, "loadRtbInterstitialAd...");
+  }
+
+  @Override
+  public void loadInterstitialAd(
+      @NonNull MediationInterstitialAdConfiguration mediationInterstitialAdConfiguration,
+      @NonNull MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback> mediationAdLoadCallback) {
+    Log.d(TAG, "loadInterstitialAd...");
+  }
+}
