+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/google/ads/mediation/vungle/VungleMediationAdapter.java
@@ -6,17 +6,15 @@ import android.os.Handler;
 import android.os.Looper;
 import android.text.TextUtils;
 import android.util.Log;
+import androidx.annotation.NonNull;
 import com.google.ads.mediation.vungle.VungleInitializer.VungleInitializationListener;
-import com.google.android.gms.ads.mediation.Adapter;
-import com.google.android.gms.ads.mediation.InitializationCompleteCallback;
+import com.google.android.gms.ads.AdError;
+import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback;
-import com.google.android.gms.ads.mediation.MediationConfiguration;
 import com.google.android.gms.ads.mediation.MediationRewardedAd;
 import com.google.android.gms.ads.mediation.MediationRewardedAdCallback;
 import com.google.android.gms.ads.mediation.MediationRewardedAdConfiguration;
-import com.google.android.gms.ads.mediation.VersionInfo;
 import com.google.android.gms.ads.rewarded.RewardItem;
-import com.vungle.mediation.BuildConfig;
 import com.vungle.mediation.VungleExtrasBuilder;
 import com.vungle.mediation.VungleManager;
 import com.vungle.warren.AdConfig;
@@ -26,24 +24,21 @@ import com.vungle.warren.Vungle;
 import com.vungle.warren.error.VungleException;
 import java.lang.ref.WeakReference;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
 
 /**
  * Mediation network adapter for Vungle.
  */
-public class VungleMediationAdapter extends Adapter
+public class VungleMediationAdapter extends VungleBaseAdapter
     implements MediationRewardedAd, LoadAdCallback, PlayAdCallback {
 
   private static final String TAG = VungleMediationAdapter.class.getSimpleName();
-  private static final String KEY_APP_ID = "appid";
 
   private AdConfig mAdConfig;
   private String mUserID;
   private String mPlacement;
-  private Handler mHandler = new Handler(Looper.getMainLooper());
+  private final Handler mHandler = new Handler(Looper.getMainLooper());
 
-  private static HashMap<String, WeakReference<VungleMediationAdapter>> mPlacementsInUse =
+  private static final HashMap<String, WeakReference<VungleMediationAdapter>> mPlacementsInUse =
       new HashMap<>();
 
   private MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
@@ -51,103 +46,9 @@ public class VungleMediationAdapter extends Adapter
   private MediationRewardedAdCallback mMediationRewardedAdCallback;
 
   @Override
-  public VersionInfo getVersionInfo() {
-    String versionString = BuildConfig.ADAPTER_VERSION;
-    String[] splits = versionString.split("\\.");
-
-    if (splits.length >= 4) {
-      int major = Integer.parseInt(splits[0]);
-      int minor = Integer.parseInt(splits[1]);
-      int micro = Integer.parseInt(splits[2]) * 100 + Integer.parseInt(splits[3]);
-      return new VersionInfo(major, minor, micro);
-    }
-
-    String logMessage =
-        String.format(
-            "Unexpected adapter version format: %s. Returning 0.0.0 for adapter version.",
-            versionString);
-    Log.w(TAG, logMessage);
-    return new VersionInfo(0, 0, 0);
-  }
-
-  @Override
-  public VersionInfo getSDKVersionInfo() {
-    String versionString = com.vungle.warren.BuildConfig.VERSION_NAME;
-    String[] splits = versionString.split("\\.");
-
-    if (splits.length >= 3) {
-      int major = Integer.parseInt(splits[0]);
-      int minor = Integer.parseInt(splits[1]);
-      int micro = Integer.parseInt(splits[2]);
-      return new VersionInfo(major, minor, micro);
-    }
-
-    String logMessage =
-        String.format(
-            "Unexpected SDK version format: %s. Returning 0.0.0 for SDK version.", versionString);
-    Log.w(TAG, logMessage);
-    return new VersionInfo(0, 0, 0);
-  }
-
-  @Override
-  public void initialize(
-      Context context,
-      final InitializationCompleteCallback initializationCompleteCallback,
-      List<MediationConfiguration> mediationConfigurations) {
-
-    if (Vungle.isInitialized()) {
-      initializationCompleteCallback.onInitializationSucceeded();
-      return;
-    }
-
-    HashSet<String> appIDs = new HashSet<>();
-    for (MediationConfiguration configuration : mediationConfigurations) {
-      Bundle serverParameters = configuration.getServerParameters();
-      String appIDFromServer = serverParameters.getString(KEY_APP_ID);
-
-      if (!TextUtils.isEmpty(appIDFromServer)) {
-        appIDs.add(appIDFromServer);
-      }
-    }
-
-    int count = appIDs.size();
-    if (count <= 0) {
-      initializationCompleteCallback.onInitializationFailed(
-          "Initialization failed: Missing or Invalid App ID.");
-      return;
-    }
-
-    String appID = appIDs.iterator().next();
-    if (count > 1) {
-      String logMessage =
-          String.format(
-              "Multiple '%s' entries found: %s. Using '%s' to initialize the Vungle SDK.",
-              KEY_APP_ID, appIDs.toString(), appID);
-      Log.w(TAG, logMessage);
-    }
-
-    VungleInitializer.getInstance()
-        .initialize(
-            appID,
-            context.getApplicationContext(),
-            new VungleInitializationListener() {
-              @Override
-              public void onInitializeSuccess() {
-                initializationCompleteCallback.onInitializationSucceeded();
-              }
-
-              @Override
-              public void onInitializeError(String errorMessage) {
-                initializationCompleteCallback.onInitializationFailed(
-                    "Initialization Failed: " + errorMessage);
-              }
-            });
-  }
-
-  @Override
   public void loadRewardedAd(
-      MediationRewardedAdConfiguration mediationRewardedAdConfiguration,
-      MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
+      @NonNull MediationRewardedAdConfiguration mediationRewardedAdConfiguration,
+      @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
           mediationAdLoadCallback) {
     mMediationAdLoadCallback = mediationAdLoadCallback;
 
@@ -162,7 +63,9 @@ public class VungleMediationAdapter extends Adapter
     if (TextUtils.isEmpty(mPlacement)) {
       String logMessage = "Failed to load ad from Vungle: Missing or invalid Placement ID.";
       Log.w(TAG, logMessage);
-      mediationAdLoadCallback.onFailure(logMessage);
+      AdError error = new AdError(AdRequest.ERROR_CODE_INVALID_REQUEST, logMessage,
+          VungleMediationAdapter.this.getClass().getName());
+      mediationAdLoadCallback.onFailure(error);
       return;
     }
 
@@ -170,7 +73,9 @@ public class VungleMediationAdapter extends Adapter
         && mPlacementsInUse.get(mPlacement).get() != null) {
       String logMessage = "Only a maximum of one ad can be loaded per placement.";
       Log.w(TAG, logMessage);
-      mediationAdLoadCallback.onFailure(logMessage);
+      AdError error = new AdError(AdRequest.ERROR_CODE_INVALID_REQUEST, logMessage,
+          VungleMediationAdapter.this.getClass().getName());
+      mediationAdLoadCallback.onFailure(error);
       return;
     }
 
@@ -178,7 +83,9 @@ public class VungleMediationAdapter extends Adapter
     if (TextUtils.isEmpty(appID)) {
       String logMessage = "Failed to load ad from Vungle: Missing or Invalid App ID.";
       Log.w(TAG, logMessage);
-      mediationAdLoadCallback.onFailure(logMessage);
+      AdError error = new AdError(AdRequest.ERROR_CODE_INVALID_REQUEST, logMessage,
+          VungleMediationAdapter.this.getClass().getName());
+      mediationAdLoadCallback.onFailure(error);
       return;
     }
 
@@ -206,19 +113,24 @@ public class VungleMediationAdapter extends Adapter
               @Override
               public void onInitializeError(String errorMessage) {
                 Log.w(TAG, errorMessage);
-                mMediationAdLoadCallback.onFailure(errorMessage);
+                AdError error = new AdError(AdRequest.ERROR_CODE_INTERNAL_ERROR,
+                    "" + errorMessage,
+                    VungleMediationAdapter.this.getClass().getName());
+                mMediationAdLoadCallback.onFailure(error);
                 mPlacementsInUse.remove(mPlacement);
               }
             });
   }
 
   @Override
-  public void showAd(Context context) {
+  public void showAd(@NonNull Context context) {
     if (Vungle.canPlayAd(mPlacement)) {
       Vungle.playAd(mPlacement, mAdConfig, VungleMediationAdapter.this);
     } else {
       if (mMediationRewardedAdCallback != null) {
-        mMediationRewardedAdCallback.onAdFailedToShow("Not ready.");
+        AdError error = new AdError(AdRequest.ERROR_CODE_NO_FILL, "Not ready.",
+            VungleMediationAdapter.this.getClass().getName());
+        mMediationRewardedAdCallback.onAdFailedToShow(error);
       }
       mPlacementsInUse.remove(mPlacement);
     }
@@ -325,13 +237,16 @@ public class VungleMediationAdapter extends Adapter
         new Runnable() {
           @Override
           public void run() {
+            AdError error = new AdError(throwable.getExceptionCode(),
+                "" + throwable.getLocalizedMessage(),
+                VungleMediationAdapter.this.getClass().getName());
             if (mMediationAdLoadCallback != null) {
               Log.w(TAG, "Failed to load ad from Vungle", throwable);
-              mMediationAdLoadCallback.onFailure(throwable.getLocalizedMessage());
+              mMediationAdLoadCallback.onFailure(error);
             }
 
             if (mMediationRewardedAdCallback != null) {
-              mMediationRewardedAdCallback.onAdFailedToShow(throwable.getLocalizedMessage());
+              mMediationRewardedAdCallback.onAdFailedToShow(error);
             }
             mPlacementsInUse.remove(placementId);
           }
@@ -347,7 +262,7 @@ public class VungleMediationAdapter extends Adapter
   /**
    * This class is used to map Vungle rewarded video ad rewards to Google Mobile Ads SDK rewards.
    */
-  private class VungleReward implements RewardItem {
+  private static class VungleReward implements RewardItem {
 
     private final String mType;
     private final int mAmount;
@@ -363,6 +278,7 @@ public class VungleMediationAdapter extends Adapter
     }
 
     @Override
+    @NonNull
     public String getType() {
       return mType;
     }
